refactor: move and setup the userrepository to account module
The account module will wrap the functionality between the server and database modules for handling
the user. The only class that is currently relevant is the UserRepository so move that into there
