+++ app/src/main/java/com/chesire/malime/services/RefreshUserWorker.kt
@@ -5,7 +5,7 @@ import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
 import com.chesire.malime.MalimeApplication
 import com.chesire.malime.server.Resource
-import com.chesire.malime.repo.UserRepository
+import com.chesire.malime.account.UserRepository
 import timber.log.Timber
 import javax.inject.Inject
 
@@ -15,7 +15,7 @@ class RefreshUserWorker(
 ) : CoroutineWorker(appContext, workerParams) {
 
     @Inject
-    lateinit var userRepo: UserRepository
+    lateinit var userRepo: com.chesire.malime.account.UserRepository
 
     init {
         // For now setup in the init block
