+++ app/src/test/java/com/chesire/malime/services/RefreshUserWorkerTests.kt
@@ -3,9 +3,7 @@ package com.chesire.malime.services
 import android.content.Context
 import androidx.work.ListenableWorker
 import androidx.work.WorkerParameters
-import com.chesire.malime.server.Resource
-import com.chesire.malime.repo.UserRepository
-import io.mockk.coEvery
+import com.chesire.malime.account.UserRepository
 import io.mockk.coVerify
 import io.mockk.every
 import io.mockk.mockk
@@ -22,7 +20,7 @@ class RefreshUserWorkerTests {
                 every { backgroundExecutor } returns mockk()
             }
         }
-        val mockUserRepo = mockk<UserRepository> {
+        val mockUserRepo = mockk<com.chesire.malime.account.UserRepository> {
             coEvery { retrieveUserId() } coAnswers { null }
         }
 
@@ -42,7 +40,7 @@ class RefreshUserWorkerTests {
                 every { backgroundExecutor } returns mockk()
             }
         }
-        val mockUserRepo = mockk<UserRepository> {
+        val mockUserRepo = mockk<com.chesire.malime.account.UserRepository> {
             coEvery { retrieveUserId() } coAnswers { 1 }
             coEvery { refreshUser() } coAnswers { Resource.Success(mockk()) }
         }
@@ -63,7 +61,7 @@ class RefreshUserWorkerTests {
                 every { backgroundExecutor } returns mockk()
             }
         }
-        val mockUserRepo = mockk<UserRepository> {
+        val mockUserRepo = mockk<com.chesire.malime.account.UserRepository> {
             coEvery { retrieveUserId() } coAnswers { 1 }
             coEvery { refreshUser() } coAnswers { Resource.Error("") }
         }
